Factorise out Camera module.
