+++ src/ios/Diagnostic.m
@@ -209,121 +209,6 @@ ABAddressBookRef _addressBook;
 }
 
 
-#pragma mark - Camera
-- (void) isCameraAvailable: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            [self sendPluginResultBool:[self isCameraPresent] && [self isCameraAuthorized] :command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) isCameraPresent: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            [self sendPluginResultBool:[self isCameraPresent] :command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) isCameraAuthorized: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            [self sendPluginResultBool:[self isCameraAuthorized] :command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) getCameraAuthorizationStatus: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            NSString* status;
-            AVAuthorizationStatus authStatus = [AVCaptureDevice authorizationStatusForMediaType:AVMediaTypeVideo];
-            
-            if(authStatus == AVAuthorizationStatusDenied || authStatus == AVAuthorizationStatusRestricted){
-                status = AUTHORIZATION_DENIED;
-            }else if(authStatus == AVAuthorizationStatusNotDetermined){
-                status = AUTHORIZATION_NOT_DETERMINED;
-            }else if(authStatus == AVAuthorizationStatusAuthorized){
-                status = AUTHORIZATION_GRANTED;
-            }
-            [self logDebug:[NSString stringWithFormat:@"Camera authorization status is: %@", status]];
-            [self sendPluginResultString:status:command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) requestCameraAuthorization: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            [AVCaptureDevice requestAccessForMediaType:AVMediaTypeVideo completionHandler:^(BOOL granted) {
-                [self sendPluginResultBool:granted :command];
-            }];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) isCameraRollAuthorized: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            [self sendPluginResultBool:[[self getCameraRollAuthorizationStatus]  isEqual: AUTHORIZATION_GRANTED] :command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) getCameraRollAuthorizationStatus: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            NSString* status = [self getCameraRollAuthorizationStatus];
-            [self logDebug:[NSString stringWithFormat:@"Camera Roll authorization status is: %@", status]];
-            [self sendPluginResultString:status:command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) requestCameraRollAuthorization: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            [PHPhotoLibrary requestAuthorization:^(PHAuthorizationStatus authStatus) {
-                NSString* status = [self getCameraRollAuthorizationStatusAsString:authStatus];
-                [self sendPluginResultString:status:command];
-            }];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
 
 #pragma mark -  Settings
 - (void) switchToSettings: (CDVInvokedUrlCommand*)command
@@ -1089,51 +974,6 @@ ABAddressBookRef _addressBook;
     return [CMPedometer respondsToSelector:@selector(isPedometerEventTrackingAvailable)] && [CMPedometer isPedometerEventTrackingAvailable];
 }
 
-
-- (BOOL) isCameraPresent
-{
-    BOOL cameraAvailable =
-    [UIImagePickerController
-     isSourceTypeAvailable:UIImagePickerControllerSourceTypeCamera];
-    if(cameraAvailable) {
-        [self logDebug:@"Camera available"];
-        return true;
-    }
-    else {
-        [self logDebug:@"Camera unavailable"];
-        return false;
-    }
-}
-
-- (BOOL) isCameraAuthorized
-{
-    AVAuthorizationStatus authStatus = [AVCaptureDevice authorizationStatusForMediaType:AVMediaTypeVideo];
-    if(authStatus == AVAuthorizationStatusAuthorized) {
-        return true;
-    } else {
-        return false;
-    }
-}
-
-- (NSString*) getCameraRollAuthorizationStatus
-{
-    PHAuthorizationStatus authStatus = [PHPhotoLibrary authorizationStatus];
-    return [self getCameraRollAuthorizationStatusAsString:authStatus];
-    
-}
-
-- (NSString*) getCameraRollAuthorizationStatusAsString: (PHAuthorizationStatus)authStatus
-{
-    NSString* status;
-    if(authStatus == PHAuthorizationStatusDenied || authStatus == PHAuthorizationStatusRestricted){
-        status = AUTHORIZATION_DENIED;
-    }else if(authStatus == PHAuthorizationStatusNotDetermined ){
-        status = AUTHORIZATION_NOT_DETERMINED;
-    }else if(authStatus == PHAuthorizationStatusAuthorized){
-        status = AUTHORIZATION_GRANTED;
-    }
-    return status;
-}
 
 #if __IPHONE_OS_VERSION_MAX_ALLOWED < __IPHONE_9_0
 - (ABAddressBookRef)addressBook {
