+++ src/ios/Diagnostic_Camera.h
@@ -1,0 +1,27 @@
+/*
+ *  Diagnostic_Camera.h
+ *  Diagnostic Plugin - Camera Module
+ *
+ *  Copyright (c) 2018 Working Edge Ltd.
+ *  Copyright (c) 2012 AVANTIC ESTUDIO DE INGENIEROS
+ */
+
+#import <Cordova/CDV.h>
+#import <Cordova/CDVPlugin.h>
+#import "Diagnostic.h"
+
+#import <AVFoundation/AVFoundation.h>
+#import <Photos/Photos.h>
+
+
+@interface Diagnostic_Camera : CDVPlugin
+
+- (void) isCameraAvailable: (CDVInvokedUrlCommand*)command;
+- (void) isCameraPresent: (CDVInvokedUrlCommand*)command;
+- (void) isCameraAuthorized: (CDVInvokedUrlCommand*)command;
+- (void) getCameraAuthorizationStatus: (CDVInvokedUrlCommand*)command;
+- (void) requestCameraAuthorization: (CDVInvokedUrlCommand*)command;
+- (void) isCameraRollAuthorized: (CDVInvokedUrlCommand*)command;
+- (void) getCameraRollAuthorizationStatus: (CDVInvokedUrlCommand*)command;
+
+@end
